requirements-dev.txt
pytest
model_mommy
factory_boy
pytest-django